Adding tests for calling pipeline endpoint with JWT.
