+++ src/main/java/gov/cdc/izgateway/xform/services/AccessControlService.java
@@ -1,32 +1,19 @@
 package gov.cdc.izgateway.xform.services;
 
 import gov.cdc.izgateway.logging.RequestContext;
-import gov.cdc.izgateway.model.IAccessControl;
 import gov.cdc.izgateway.security.AccessControlRegistry;
-import gov.cdc.izgateway.service.IAccessControlService;
 import gov.cdc.izgateway.xform.model.AccessControl;
 import gov.cdc.izgateway.xform.repository.XformRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestMethod;
-
 import java.util.*;
 
-/**
- * The AccessControlService is a placeholder for the IAccessControlService interface.  We will be implementing
- * a simpler repository than the existing DB repository.
- * This class contains no-ops for some methods.
- * Ticket to track this: https://support.izgateway.org/browse/IGDD-1664
- */
 @Slf4j
 @Service
-public class AccessControlService  extends GenericService<AccessControl> implements IAccessControlService {
+public class AccessControlService extends GenericService<AccessControl> {
     private final AccessControlRegistry registry;
-
-    @Value("${server.hostname:dev.izgateway.org}")
-    private String serverName;
 
     @Autowired
     public AccessControlService(XformRepository<AccessControl> repo, AccessControlRegistry registry) {
@@ -40,20 +27,13 @@ public class AccessControlService  extends GenericService<AccessControl> impleme
         return roles;
     }
 
-    // TODO: PCahill - discuss ... we don't need the user parameter because the request context has the current user
-    //       and their roles. Should we update the interface to allow a checkAccess method that doesn't require
-    //       a user parameter?
-    @Override
-    // TODO PCAHILL - from Code Review - we don't need to call this, instead we can call whatever it is that we need.
-    // don't make any changes to the overridden method.
-    public Boolean checkAccess(String user, String method, String path) {
+    public Boolean checkAccess(String method, String path) {
         List<String> allowedRoles = getAllowedRoles(RequestMethod.valueOf(method), path);
 
         // If RequestContext.getRoles() has one role that matches the roles list, return true
         return RequestContext.getPrincipal().getRoles().stream().anyMatch(allowedRoles::contains);
     }
 
-    @Override
     public Map<String, TreeSet<String>> getUserRoles() {
         Map<String, TreeSet<String>> userRoleMap = new HashMap<>();
         for (AccessControl ac : repo.getEntitySet()) {
@@ -62,69 +42,4 @@ public class AccessControlService  extends GenericService<AccessControl> impleme
 
         return userRoleMap;
     }
-
-
-    @Override
-    public String getServerName() {
-        return serverName;
-    }
-
-    @Override
-    public void refresh() {
-    }
-
-    @Override
-    public Map<String, Map<String, Boolean>> getAllowedUsersByGroup() {
-        return Map.of();
-    }
-
-    @Override
-    public Map<String, Map<String, Boolean>> getAllowedRoutesByEvent() {
-        return Map.of();
-    }
-
-    @Override
-    public boolean isUserInRole(String user, String role) {
-        return false;
-    }
-
-    @Override
-    public boolean isUserBlacklisted(String user) {
-        return false;
-    }
-
-    @Override
-    public Map<String, Boolean> getEventMap(String event) {
-        return Map.of();
-    }
-
-    @Override
-    public Set<String> getEventTypes() {
-        return Set.of();
-    }
-
-    @Override
-    public boolean isRouteAllowed(String route, String event) {
-        return false;
-    }
-
-    @Override
-    public void setServerName(String serverName) {
-    }
-
-    @Override
-    public boolean isMemberOf(String user, String group) {
-        return false;
-    }
-
-    @Override
-    public IAccessControl removeUserFromBlacklist(String user) {
-        return null;
-    }
-
-    @Override
-    public IAccessControl addUserToBlacklist(String user) {
-        return null;
-    }
-
 }
