+++ testing/configuration/access-control.json
@@ -4,7 +4,17 @@
     "userId": "8c942339-e10d-418f-8898-d8e73388198c",
     "active": true,
     "roles": [
-      "xform-sender"
+      "admin",
+      "xform-sender",
+      "pipeline-reader",
+      "pipeline-writer",
+      "pipeline-deleter",
+      "organization-reader",
+      "organization-writer",
+      "organization-deleter",
+      "solution-reader",
+      "solution-writer",
+      "solution-deleter"
     ]
   },
   {
@@ -12,8 +22,17 @@
     "userId": "f1d0616c-ec7c-4070-b43a-9e7295d45d21",
     "active": true,
     "roles": [
+      "admin",
       "xform-sender",
-      "admin"
+      "pipeline-reader",
+      "pipeline-writer",
+      "pipeline-deleter",
+      "organization-reader",
+      "organization-writer",
+      "organization-deleter",
+      "solution-reader",
+      "solution-writer",
+      "solution-deleter"
     ]
   },
   {
@@ -21,7 +40,17 @@
     "userId": "61d7e381-2807-4d0f-8e6a-8a0df182d70d",
     "active": true,
     "roles": [
-      "admin"
+      "admin",
+      "xform-sender",
+      "pipeline-reader",
+      "pipeline-writer",
+      "pipeline-deleter",
+      "organization-reader",
+      "organization-writer",
+      "organization-deleter",
+      "solution-reader",
+      "solution-writer",
+      "solution-deleter"
     ]
   }
 ]
